src/backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.27.0

# Data Validation
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
databases[postgresql]>=0.8.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Monitoring & Metrics
prometheus-fastapi-instrumentator>=6.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
